// src/lib/spacetime/generated/types.ts
export const Resident = __t.object("Resident", {
  id: __t.string(),
  roomId: __t.string(),
  branchId: __t.string(),
  name: __t.string(),
  role: __t.string(),
  appearanceId: __t.u32(),
  x: __t.f32(),
  y: __t.f32(),
  state: __t.string(),
  revision: __t.u32(),
});
export type Resident = __Infer<typeof Resident>;